+++ src/app/components/settings-dialogs/settings-dialogs.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { BhPolizaServiceService } from 'src/app/services/bh-poliza-service.service';
-import { Choice, Message } from '../bharatrpatil-chat/bharatrpatil-chat.component';
+import { Message } from './../../Message';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-settings-dialogs',
@@ -46,11 +47,30 @@ export class SettingsDialogsComponent implements OnInit {
     this.DialogForm.controls["url"].setValue(Data.url);
     this.DialogForm.controls["format"].setValue(Data.format);
     this.EventValueDialog = "Actualizar";
+    this.onChangeType(null);
   }
 
   deleteDataDialog(id) {
-    this.polizaService.deleteDialog(id).subscribe(() => {
-      this.getDataDialog();
+    Swal.fire({
+      title: '¿Estás seguro que deseas eliminar este elemento?',
+      text: "No lo podrás recuperar después.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar.'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.polizaService.deleteDialog(id).subscribe(() => {
+          this.getDataDialog();
+          this.resetFormDialog();
+          Swal.fire(
+            'Eliminado!',
+            'Se eliminó el elemento correctamente',
+            'success'
+          );
+        })
+      }
     })
   }
 
@@ -84,6 +104,26 @@ export class SettingsDialogsComponent implements OnInit {
     this.polizaService.putDialog(this.DialogForm.value.id, this.DialogForm.value).subscribe((data: any) => {
       this.resetFormDialog();
     })
+  }
+
+  onChangeType(event: any){
+    if (this.DialogForm.value.type === "message"){
+      this.typeSelectChange(this.DialogForm, [], ['url','previousQuestion','format']);
+    } else if (this.DialogForm.value.type === "input"){
+      this.typeSelectChange(this.DialogForm, ['url','previousQuestion'], ['format']);
+    } else if (this.DialogForm.value.type === "choices") {
+      this.typeSelectChange(this.DialogForm, ['previousQuestion'], ['url','format']);
+    }
+  }
+
+  typeSelectChange(form: FormGroup, enableFields: string[], disableFields: string[]){
+    for (const efield of enableFields) {
+      form.controls[efield].enable()
+    }
+    for (const dfield of disableFields) {
+      form.controls[dfield].disable()
+      form.controls[dfield].reset()
+    }
   }
 
 }
